requirements.txt
pycryptodomex
PyYAML
pywin32
resend

